+++ src/main/java/cech12/woodenshears/api/item/WoodenShearsItems.java
@@ -1,9 +1,0 @@
-package cech12.woodenshears.api.item;
-
-import net.minecraft.world.item.Item;
-
-public class WoodenShearsItems {
-
-    public static Item WOODEN_SHEARS;
-
-}