+++ CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Forge Recommended Versioning](https://mcforge.readthedocs.io/en/latest/conventions/versioning/).
 
+## [1.18.2-1.2.0.0] - 2022-05-03
+### Changed
+- Update mod to Forge 1.18.2-40.1.2 (fix interaction with entities with both hands)
+
+### Added
+- server config option to define the durability of wooden shears
+
 ## [1.18.1-1.1.1.0] - 2022-01-10
 ### Changed
 - Update mod to Forge 1.18.1-39.0.16